factors out test of generation
