+++ src/generation/parsed_to_render.rs
@@ -47,28 +47,28 @@ pub enum EventType {
 
 #[derive(Debug, Clone, Eq, Ord, PartialEq, PartialOrd)]
 pub struct TimedOp {
-    t: Rational64,
-    event_type: EventType,
-    voice: usize,
-    event: usize,
-    fm: Rational64,
-    fa: Rational64,
-    pm: Rational64,
-    pa: Rational64,
-    g: Rational64,
-    l: Rational64,
+    pub t: Rational64,
+    pub event_type: EventType,
+    pub voice: usize,
+    pub event: usize,
+    pub fm: Rational64,
+    pub fa: Rational64,
+    pub pm: Rational64,
+    pub pa: Rational64,
+    pub g: Rational64,
+    pub l: Rational64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Op4D {
-    t: f64,
-    event_type: EventType,
-    voice: usize,
-    event: usize,
-    x: f64,
-    y: f64,
-    z: f64,
-    l: f64,
+    pub t: f64,
+    pub event_type: EventType,
+    pub voice: usize,
+    pub event: usize,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+    pub l: f64,
 }
 
 impl TimedOp {
@@ -116,11 +116,11 @@ fn point_op_to_timed_op(
     (on, off)
 }
 
-fn vec_timed_op_to_vec_op4d(timed_ops: Vec<TimedOp>, basis: &Basis) -> Vec<Op4D> {
+pub fn vec_timed_op_to_vec_op4d(timed_ops: Vec<TimedOp>, basis: &Basis) -> Vec<Op4D> {
     timed_ops.iter().map(|t_op| t_op.to_op_4d(&basis)).collect()
 }
 
-fn composition_to_vec_timed_op(composition: &NormalForm, table: &OpOrNfTable) -> Vec<TimedOp> {
+pub fn composition_to_vec_timed_op(composition: &NormalForm, table: &OpOrNfTable) -> Vec<TimedOp> {
     let mut normal_form = NormalForm::init();
 
     println!("Generating Composition \n");
@@ -212,216 +212,8 @@ fn sort_vecs(vec_wav: &mut Vec<StereoWaveform>) {
     vec_wav.sort_unstable_by(|a, b| b.l_buffer.len().cmp(&a.l_buffer.len()));
 }
 
-fn sum_vec(a: &mut Vec<f64>, b: &[f64]) {
+pub fn sum_vec(a: &mut Vec<f64>, b: &[f64]) {
     for (ai, bi) in a.iter_mut().zip(b) {
         *ai += *bi;
-    }
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use socool_ast::{Op::*, OpOrNf::*};
-
-    #[test]
-    fn render_equal() {
-        let mut a = vec![1.0, 2.0, 3.0];
-        let b = vec![1.0, 2.0, 3.0];
-        sum_vec(&mut a, &b[..]);
-        let expected = [2.0, 4.0, 6.0];
-        assert_eq!(a, expected);
-    }
-
-    #[test]
-    fn render_left() {
-        let mut a = vec![1.0, 2.0, 3.0, 2.0];
-        let b = vec![1.0, 2.0, 3.0];
-        sum_vec(&mut a, &b[..]);
-        let expected = [2.0, 4.0, 6.0, 2.0];
-        assert_eq!(a, expected);
-    }
-
-    #[test]
-    fn to_vec_timed_op_test() {
-        let mut normal_form = NormalForm::init();
-        let pt = OpOrNfTable::new();
-
-        Overlay {
-            operations: vec![
-                Op(Sequence {
-                    operations: vec![
-                        Op(PanA {
-                            a: Rational64::new(1, 2),
-                        }),
-                        Op(TransposeM {
-                            m: Rational64::new(2, 1),
-                        }),
-                        Op(Gain {
-                            m: Rational64::new(1, 2),
-                        }),
-                        Op(Length {
-                            m: Rational64::new(2, 1),
-                        }),
-                    ],
-                }),
-                Op(Sequence {
-                    operations: vec![Op(Length {
-                        m: Rational64::new(5, 1),
-                    })],
-                }),
-            ],
-        }
-        .apply_to_normal_form(&mut normal_form, &pt);
-
-        let timed_ops = composition_to_vec_timed_op(&normal_form, &pt);
-
-        let op = TimedOp {
-            fm: Rational64::new(1, 1),
-            fa: Rational64::new(0, 1),
-            pm: Rational64::new(1, 1),
-            pa: Rational64::new(0, 1),
-            g: Rational64::new(1, 1),
-            l: Rational64::new(1, 1),
-            t: Rational64::new(0, 1),
-            event_type: EventType::On,
-            voice: 0,
-            event: 0,
-        };
-
-        assert_eq!(
-            timed_ops,
-            vec![
-                TimedOp {
-                    pa: Rational64::new(1, 2),
-                    event_type: EventType::On,
-                    ..op
-                },
-                TimedOp {
-                    event_type: EventType::On,
-                    l: Rational64::new(5, 1),
-                    voice: 1,
-                    ..op
-                },
-                TimedOp {
-                    pa: Rational64::new(1, 2),
-                    t: Rational64::new(1, 1),
-                    event_type: EventType::Off,
-                    ..op
-                },
-                TimedOp {
-                    fm: Rational64::new(2, 1),
-                    t: Rational64::new(1, 1),
-                    event_type: EventType::On,
-                    event: 1,
-                    ..op
-                },
-                TimedOp {
-                    fm: Rational64::new(2, 1),
-                    t: Rational64::new(2, 1),
-                    event_type: EventType::Off,
-                    event: 1,
-                    ..op
-                },
-                TimedOp {
-                    g: Rational64::new(1, 2),
-                    t: Rational64::new(2, 1),
-                    event_type: EventType::On,
-                    event: 2,
-                    ..op
-                },
-                TimedOp {
-                    g: Rational64::new(1, 2),
-                    t: Rational64::new(3, 1),
-                    event_type: EventType::Off,
-                    event: 2,
-                    ..op
-                },
-                TimedOp {
-                    t: Rational64::new(3, 1),
-                    l: Rational64::new(2, 1),
-                    event_type: EventType::On,
-                    event: 3,
-                    ..op
-                },
-                TimedOp {
-                    t: Rational64::new(5, 1),
-                    l: Rational64::new(2, 1),
-                    event_type: EventType::Off,
-                    event: 3,
-                    ..op
-                },
-                TimedOp {
-                    t: Rational64::new(5, 1),
-                    l: Rational64::new(5, 1),
-                    event_type: EventType::Off,
-                    voice: 1,
-                    ..op
-                },
-            ]
-        );
-    }
-
-    #[test]
-    fn to_vec_op4d_test() {
-        let basis = Basis {
-            f: 100.0,
-            g: 1.0,
-            p: 0.0,
-            l: 1.0,
-            a: 44100.0,
-            d: 44100.0,
-        };
-
-        let op = TimedOp {
-            fm: Rational64::new(2, 1),
-            fa: Rational64::new(0, 1),
-            pm: Rational64::new(1, 1),
-            pa: Rational64::new(1, 2),
-            g: Rational64::new(1, 2),
-            t: Rational64::new(0, 1),
-            l: Rational64::new(1, 1),
-            event_type: EventType::On,
-            voice: 0,
-            event: 0,
-        };
-
-        let vec_timed_op = vec![
-            TimedOp {
-                event_type: EventType::On,
-                l: Rational64::new(3, 2),
-                ..op
-            },
-            TimedOp {
-                event_type: EventType::Off,
-                l: Rational64::new(3, 2),
-                t: Rational64::new(3, 2),
-                ..op
-            },
-        ];
-
-        let result = vec_timed_op_to_vec_op4d(vec_timed_op, &basis);
-        let expected = vec![
-            Op4D {
-                t: 0.0,
-                l: 1.5,
-                event_type: EventType::On,
-                voice: 0,
-                event: 0,
-                x: 0.5,
-                y: 200.0,
-                z: 0.5,
-            },
-            Op4D {
-                t: 1.5,
-                l: 1.5,
-                event_type: EventType::Off,
-                voice: 0,
-                event: 0,
-                x: 0.5,
-                y: 200.0,
-                z: 0.5,
-            },
-        ];
-        assert_eq!(result, expected);
     }
 }
