+++ src/portaudio/duplex.rs
@@ -38,7 +38,7 @@ fn process_result(result: &mut DetectionResult) -> Basis {
     }
 }
 
-pub fn setup_output(
+pub fn duplex_setup(
     parsed_composition: Vec<Vec<PointOp>>,
     ref pa: &pa::PortAudio,
 ) -> Result<pa::Stream<pa::NonBlocking, pa::Duplex<f32, f32>>, pa::Error> {
